+++ internal/core/application/dto/product.go
@@ -4,14 +4,14 @@ import "io"
 
 type CreateProduct struct {
 	Name      string    `json:"name"`
-	Price     float64   `json:"price"`
+	Price     int       `json:"price"`
 	Quantity  int       `json:"quantity"`
 	Image     io.Reader `json:"image"`
 	ImageName string    `json:"image_name"`
 }
 
 type UpdatePrice struct {
-	Amount float64 `json:"amount"`
+	Amount int `json:"amount"`
 }
 
 type UpdateQuantity struct {
